+++ installers/ansible/roles/pgo-operator/files/pgo-configs/pgrestore-job.json
@@ -31,6 +31,9 @@
                 ],
                 "securityContext": {{.SecurityContext}},
                 "serviceAccountName": "pgo-default",
+                {{ if .Tolerations }}
+                "tolerations": {{ .Tolerations }},
+                {{ end }}
                 "containers": [
                     {
                         "name": "pgrestore",
