+++ internal/operator/cluster/clusterlogic.go
@@ -347,7 +347,7 @@ func getClusterDeploymentFields(clientset kubernetes.Interface,
 		ReplicationTLSSecret:      cl.Spec.TLS.ReplicationTLSSecret,
 		CASecret:                  cl.Spec.TLS.CASecret,
 		Standby:                   cl.Spec.Standby,
-		Tolerations:               operator.GetTolerations(cl.Spec.Tolerations),
+		Tolerations:               util.GetTolerations(cl.Spec.Tolerations),
 	}
 
 	return deploymentFields
@@ -491,8 +491,8 @@ func scaleReplicaCreateDeployment(clientset kubernetes.Interface,
 		// Give precedence to the tolerations defined on the replica spec, otherwise
 		// take any tolerations defined on the cluster spec
 		Tolerations: util.GetValueOrDefault(
-			operator.GetTolerations(replica.Spec.Tolerations),
-			operator.GetTolerations(cluster.Spec.Tolerations)),
+			util.GetTolerations(replica.Spec.Tolerations),
+			util.GetTolerations(cluster.Spec.Tolerations)),
 	}
 
 	switch replica.Spec.ReplicaStorage.StorageType {
